Create a new util function `run_tool_and_install_it_if_not_present`
